Refactor and add incentives tests
