+++ server/views/pages/profile.njk
@@ -46,7 +46,7 @@
     </div>
 
     <div class="card-background grey">
-      <div class="govuk-width-container" data-test="incentive-container">
+      <div class="govuk-width-container" data-test="incentives-container">
         <h2 class="govuk-heading-m">Incentives (IEP)</h2>
         {% if data.incentivesData.error %}
           <p class="govuk-!-font-size-24" data-test="incentives-error">We are not able to show your incentive level at this time. <a href="{{ params.returnUrl }}" class="govuk-link">Try again</a> or <a href="{{ incentivesSummary.link }}" class="govuk-link">find out more about incentive levels</a>.</p>
@@ -56,7 +56,7 @@
               {{ basicCard({ id: 'currentLevel', heading: 'My current level:', content: data.incentivesData.iepLevel }) }}
             </li>
           </ul>
-          <a href= '{{ data.incentivesReadMoreURL }}' rel="noreferrer noopener" target="_blank" class="govuk-!-font-size-24 card-title govuk-link govuk-!-font-weight-regular govuk-link--no-visited-state"data-test="incentivesLink">
+          <a href= '{{ data.incentivesReadMoreURL }}' rel="noreferrer noopener" target="_blank" class="govuk-!-font-size-24 card-title govuk-link govuk-!-font-weight-regular govuk-link--no-visited-state" data-test="incentivesLink">
             <span>Read more about incentives (IEP)</span>
           </a>
         {% endif %}
